feat: Replace throw by assert
Closes #41

Instead of throwing logical_error, use assert.

A special macro assert_message is added. It is a wrapper around assert
from assert.h allowing to have a message associated to the assert.

The runtime_error is kept as it catch a bug in the library and not a
developer misuse